Edited readme and more mech file renaming
