+++ jaxley_mech/channels/hodgkin52.py
