+++ jaxley_mech/synapses/__init__.py
@@ -1,4 +1,4 @@
-from .dms98 import AMPA, NMDA, GABAa, GABAb
+from .destexhe98 import AMPA, NMDA, GABAa, GABAb
 from .gap_junction import GapJunction
 from .ribbon import RibbonSynapse
 from .ribbon_mGluR6 import Ribbon_mGluR6
